pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dingtalk-ai-robot"
version = "0.1.0"
description = "钉钉智能机器人，支持AI问答、知识库检索、JIRA管理和服务器维护"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "dingtalk-stream>=2.1.0",
    "alibabacloud-dingtalk-robot>=2.1.16",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "autogen-agentchat>=0.5.7",
    "autogen-ext[openai]>=0.5.7",
    "python-jose>=3.3.0",
    "chromadb>=0.4.22",
    "paramiko>=3.3.1",
    "loguru>=0.7.2",
    "schedule>=1.2.1",
    "jira>=3.5.2",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.2",
    "black>=23.12.0",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "ruff>=0.1.11",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "SIM", "RUF"]
